+++ logfire/_internal/cli/__init__.py
@@ -7,6 +7,7 @@ import functools
 import logging
 import platform
 import sys
+import warnings
 import webbrowser
 from collections.abc import Sequence
 from operator import itemgetter
@@ -18,6 +19,7 @@ import requests
 from opentelemetry import trace
 from rich.console import Console
 
+from logfire._internal.cli.prompt import parse_prompt
 from logfire.exceptions import LogfireConfigError
 from logfire.propagate import ContextCarrier, get_context
 
@@ -225,6 +227,13 @@ def parse_create_new_project(args: argparse.Namespace) -> None:
     sys.stderr.write(f'Project created successfully. You will be able to view it at: {credentials.project_url}\n')
 
 
+def parse_create_read_token(args: argparse.Namespace) -> None:
+    """Create a read token for a project."""
+    client = LogfireClient.from_url(args.logfire_url)
+    response = client.create_read_token(args.organization, args.project)
+    sys.stdout.write(response['token'] + '\n')
+
+
 def parse_use_project(args: argparse.Namespace) -> None:
     """Use an existing project."""
     data_dir = Path(args.data_dir)
@@ -329,6 +338,27 @@ class SplitArgs(argparse.Action):
         setattr(namespace, self.dest, namespace_value + list(values or []))
 
 
+class OrgProjectAction(argparse.Action):
+    def __call__(
+        self,
+        parser: argparse.ArgumentParser,
+        namespace: argparse.Namespace,
+        values: str | Sequence[Any] | None,
+        option_string: str | None = None,
+    ):
+        if isinstance(values, str) and '/' in values:
+            try:
+                organization, project = values.split('/')
+                if not organization or not project:
+                    parser.error(f'Invalid format: {values}. Expected <org>/<project>')
+                setattr(namespace, 'organization', organization)
+                setattr(namespace, self.dest, project)
+            except ValueError:
+                parser.error(f'Invalid format: {values}. Expected <org>/<project>')
+        else:
+            parser.error(f'Invalid format: {values}. Expected <org>/<project>')
+
+
 def _main(args: list[str] | None = None) -> None:
     parser = argparse.ArgumentParser(
         prog='logfire',
@@ -340,6 +370,7 @@ def _main(args: list[str] | None = None) -> None:
     global_opts = parser.add_argument_group(title='global options')
     url_or_region_grp = global_opts.add_mutually_exclusive_group()
     url_or_region_grp.add_argument('--logfire-url', help=argparse.SUPPRESS)
+    url_or_region_grp.add_argument('--base-url', help=argparse.SUPPRESS)
     url_or_region_grp.add_argument('--region', choices=REGIONS, help='the region to use')
     parser.set_defaults(func=lambda _: parser.print_help())  # type: ignore
     subparsers = parser.add_subparsers(title='commands', metavar='')
@@ -383,6 +414,25 @@ def _main(args: list[str] | None = None) -> None:
     cmd_projects_use.add_argument('--data-dir', default='.logfire')
     cmd_projects_use.set_defaults(func=parse_use_project)
 
+    cmd_read_tokens = subparsers.add_parser('read-tokens', help='Manage read tokens for a project')
+    cmd_read_tokens.add_argument('--project', action=OrgProjectAction, help='project in the format <org>/<project>')
+    cmd_read_tokens.set_defaults(func=lambda _: cmd_read_tokens.print_help())  # type: ignore
+    read_tokens_subparsers = cmd_read_tokens.add_subparsers()
+
+    # With this command you can do:
+    # claude mcp add logfire -e LOGFIRE_READ_TOKEN=$(logfire read-tokens --project kludex/potato create) -- uvx logfire-mcp@latest
+    cmd_read_tokens_create = read_tokens_subparsers.add_parser('create', help=parse_create_read_token.__doc__)
+    cmd_read_tokens_create.set_defaults(func=parse_create_read_token)
+
+    cmd_prompt = subparsers.add_parser('prompt', help=parse_prompt.__doc__)
+    agent_code_group = cmd_prompt.add_argument_group(title='code agentic specific options')
+    claude_or_codex_group = agent_code_group.add_mutually_exclusive_group()
+    claude_or_codex_group.add_argument('--claude', action='store_true', help='verify the Claude Code setup')
+    claude_or_codex_group.add_argument('--codex', action='store_true', help='verify the Cursor setup')
+    cmd_prompt.add_argument('--project', action=OrgProjectAction, help='project in the format <org>/<project>')
+    cmd_prompt.add_argument('issue', nargs='?', help='the issue to get a prompt for')
+    cmd_prompt.set_defaults(func=parse_prompt)
+
     cmd_info = subparsers.add_parser('info', help=parse_info.__doc__)
     cmd_info.set_defaults(func=parse_info)
 
@@ -402,6 +452,15 @@ def _main(args: list[str] | None = None) -> None:
         namespace.script_and_args = unknown_args + (namespace.script_and_args or [])
     else:
         namespace = parser.parse_args(args)
+
+    if namespace.logfire_url:
+        warnings.warn(
+            'The `--logfire-url` argument is deprecated. Use `--base-url` instead.',
+            DeprecationWarning,
+            stacklevel=2,
+        )
+
+    namespace.logfire_url = namespace.logfire_url or namespace.base_url
     namespace.logfire_url = _get_logfire_url(namespace.logfire_url, namespace.region)
 
     trace.set_tracer_provider(tracer_provider=SDKTracerProvider())
@@ -410,18 +469,17 @@ def _main(args: list[str] | None = None) -> None:
     def log_trace_id(response: requests.Response, context: ContextCarrier, *args: Any, **kwargs: Any) -> None:
         logger.debug('context=%s url=%s', context, response.url)
 
-    with tracer.start_as_current_span('logfire._internal.cli'):
-        if namespace.version:
-            version_callback()
-        elif namespace.func == parse_info:
+    if namespace.version:
+        version_callback()
+    elif namespace.func in (parse_info, parse_run):
+        namespace.func(namespace)
+    else:
+        with tracer.start_as_current_span('logfire._internal.cli'), requests.Session() as session:
+            context = get_context()
+            session.hooks = {'response': functools.partial(log_trace_id, context=context)}
+            session.headers.update(context)
+            namespace._session = session
             namespace.func(namespace)
-        else:
-            with requests.Session() as session:
-                context = get_context()
-                session.hooks = {'response': functools.partial(log_trace_id, context=context)}
-                session.headers.update(context)
-                namespace._session = session
-                namespace.func(namespace)
 
 
 def main(args: list[str] | None = None) -> None: