+++ logfire/_internal/json_formatter.py
@@ -68,9 +68,9 @@ class JsonArgsValueFormatter:
         if schema is not None:
             if 'type' in schema:
                 if (data_type := schema.get('x-python-datatype')) is None:
-                    if schema['type'] == 'object':
+                    if schema['type'] == 'object' and isinstance(value, dict):
                         self._format_items('{', ': ', '}', True, indent_current, value, None)
-                    elif schema['type'] == 'array':
+                    elif schema['type'] == 'array' and isinstance(value, list):
                         self._format_list_like('[', ']', indent_current, value, schema)
                     else:
                         # e.g. {'type': 'string', 'format': 'date-time'}