+++ logfire/_internal/main.py
@@ -1353,7 +1353,7 @@ class Logfire:
         self,
         client: httpx.Client | httpx.AsyncClient | None = None,
         *,
-        capture_all: bool = False,
+        capture_all: bool | None = None,
         capture_headers: bool = False,
         capture_request_body: bool = False,
         capture_response_body: bool = False,
@@ -1375,6 +1375,7 @@ class Logfire:
             client: The `httpx.Client` or `httpx.AsyncClient` instance to instrument.
                 If `None`, the default, all clients will be instrumented.
             capture_all: Set to `True` to capture all HTTP headers, request and response bodies.
+                By default checks the environment variable `LOGFIRE_HTTPX_CAPTURE_ALL`.
             capture_headers: Set to `True` to capture all HTTP headers.
 
                 If you don't want to capture all headers, you can customize the headers captured. See the