+++ CHANGELOG.md
@@ -1,5 +1,48 @@
 # Release Notes
 
+## [v4.3.6] (2025-08-26)
+
+* Add specific code agent options for `logfire prompt` by @Kludex in [#1350](https://github.com/pydantic/logfire/pull/1350)
+
+## [v4.3.5] (2025-08-22)
+
+* Add `--base-url` to CLI by @Kludex in [#1345](https://github.com/pydantic/logfire/pull/1345)
+* Don't require stream kwarg in OpenAI methods by @alexmojaki in [#1348](https://github.com/pydantic/logfire/pull/1348)
+
+## [v4.3.4] (2025-08-21)
+
+* `logfire.instrument_google_genai()` fixes for `response_schema` and elided content by @alexmojaki in [#1342](https://github.com/pydantic/logfire/pull/1342)
+* Fix `enable_commenter` and `commenter_options` args of `instrument_sqlalchemy` by @alexmojaki in [#1335](https://github.com/pydantic/logfire/pull/1335)
+* Fix LangChain instrumentation by @alexmojaki in [#1319](https://github.com/pydantic/logfire/pull/1319)
+* Suppress `urllib3` logs while checking token by @alexmojaki in [#1341](https://github.com/pydantic/logfire/pull/1341)
+
+## [v4.3.3] (2025-08-13)
+
+* Ensure `logfire_api.LogfireSpan.context` and other attrs are None by @alexmojaki in [#1317](https://github.com/pydantic/logfire/pull/1317)
+
+## [v4.3.2] (2025-08-13)
+
+* Able to see output from token selection by @Kludex in [#1312](https://github.com/pydantic/logfire/pull/1312)
+* Remove trailing `/` from prompts call by @Kludex in [#1314](https://github.com/pydantic/logfire/pull/1314)
+* Change get read token info endpoint by @hramezani in [#1309](https://github.com/pydantic/logfire/pull/1309)
+
+## [v4.3.1] (2025-08-12)
+
+* Fix `logfire_api` requiring OpenTelemetry SDK by @alexmojaki in [#1310](https://github.com/pydantic/logfire/pull/1310)
+
+## [v4.3.0] (2025-08-12)
+
+* Closes logfire_api.LogfireSpan should support ReadableSpan attributes #989 by @RichardMarto in [#1288](https://github.com/pydantic/logfire/pull/1288)
+* Fix `test_instrument_google_genai` for specific versions by @alexmojaki in [#1303](https://github.com/pydantic/logfire/pull/1303)
+* Create read tokens via CLI by @Kludex in [#1306](https://github.com/pydantic/logfire/pull/1306)
+* Retrieve the prompt to fix issues via CLI by @Kludex in [#1307](https://github.com/pydantic/logfire/pull/1307)
+
+## [v4.2.0] (2025-08-06)
+
+* Handle wrong JSON schema in console exporter by @alexmojaki in [#1294](https://github.com/pydantic/logfire/pull/1294)
+* Move creation of span down on cli by @Kludex in [#1298](https://github.com/pydantic/logfire/pull/1298)
+* Allow to set `instrument_httpx(capture_all=True)` via env var by @Kludex in [#1295](https://github.com/pydantic/logfire/pull/1295)
+
 ## [v4.1.0] (2025-08-04)
 
 * Add `min_level` argument to `logfire.configure` by @alexmojaki in [#1265](https://github.com/pydantic/logfire/pull/1265)
@@ -820,3 +863,11 @@ First release from new repo!
 [v4.0.0]: https://github.com/pydantic/logfire/compare/v3.25.0...v4.0.0
 [v4.0.1]: https://github.com/pydantic/logfire/compare/v4.0.0...v4.0.1
 [v4.1.0]: https://github.com/pydantic/logfire/compare/v4.0.1...v4.1.0
+[v4.2.0]: https://github.com/pydantic/logfire/compare/v4.1.0...v4.2.0
+[v4.3.0]: https://github.com/pydantic/logfire/compare/v4.2.0...v4.3.0
+[v4.3.1]: https://github.com/pydantic/logfire/compare/v4.3.0...v4.3.1
+[v4.3.2]: https://github.com/pydantic/logfire/compare/v4.3.1...v4.3.2
+[v4.3.3]: https://github.com/pydantic/logfire/compare/v4.3.2...v4.3.3
+[v4.3.4]: https://github.com/pydantic/logfire/compare/v4.3.3...v4.3.4
+[v4.3.5]: https://github.com/pydantic/logfire/compare/v4.3.4...v4.3.5
+[v4.3.6]: https://github.com/pydantic/logfire/compare/v4.3.5...v4.3.6